Remove go-fed
