--- api/activitypub/activitypub.go
@@ -1,8 +1,0 @@
-package activitypub
-
-import "github.com/go-fed/activity/pub"
-
-// NewPubber creates a new activity.Pubber
-func NewPubber() pub.Pubber {
-	panic("not implemented")
-}
