--- api/activitypub/clock.go
@@ -1,10 +1,0 @@
-package activitypub
-
-import "time"
-
-// Implements pub.Clock
-type clock struct{}
-
-func (c clock) Now() {
-	time.Now()
-}
