--- api/activitypub/http_client.go
@@ -1,9 +1,0 @@
-package activitypub
-
-import "net/http"
-
-type httpClient struct{}
-
-func (h httpClient) Do(req *http.Request) (*http.Response, error) {
-	panic("not implemented")
-}
