+++ api/controllers/activitypub_test.go
@@ -1,4 +1,4 @@
-package activitypub
+package controllers
 
 import (
 	"context"
@@ -26,7 +26,7 @@ func TestOwnsCanConformToTheConfig(t *testing.T) {
 
 	u, _ := url.Parse("https://bestjeanist.com:8080")
 
-	doesOwn := application.Owns(application{}, context.TODO(), u)
+	doesOwn := Owns(context.TODO(), u)
 	assert.True(t, doesOwn, "Owns should returns true if the url matches the config")
 }
 
@@ -38,7 +38,7 @@ func TestOwnsFailsWithWrongHostname(t *testing.T) {
 
 	u, _ := url.Parse("https://goodboys.com:8080")
 
-	doesOwn := application.Owns(application{}, context.TODO(), u)
+	doesOwn := Owns(context.TODO(), u)
 	assert.False(t, doesOwn, "Owns returns false if the url doesnt match the config")
 }
 
@@ -84,7 +84,7 @@ func TestHasFailsIfNothingExists(t *testing.T) {
 	var base = "https://podocasto.com:8080"
 	u, _ := url.Parse(base + "/activity/group/sam/inbox")
 
-	val, err := application.Has(application{}, context.TODO(), u)
+	val, err := Has(context.TODO(), u)
 	assert.Nil(t, err)
 	assert.False(t, val)
 }
@@ -103,7 +103,7 @@ func TestHasPassesIfSomethingExists(t *testing.T) {
 	var base = "https://podocasto.com:8080"
 	u, _ := url.Parse(base + "/activity/group/" + slug + "/inbox")
 
-	val, err := application.Has(application{}, context.TODO(), u)
+	val, err := Has(context.TODO(), u)
 	assert.Nil(t, err)
 	assert.True(t, val)
 }
